+++ src/types/class.js
@@ -921,8 +921,7 @@ flair.Class = (arg1, arg2, arg3, arg4) => {
         interfaces: interfaces,
         name: className,
         type: 'class',
-        namespace: '',
-        assembly: null,
+        package: null,
         singleInstance: () => { return null; },
         isSingleton: () => { return false; },
         isSealed: () => { return false; },
@@ -959,6 +958,9 @@ flair.Class = (arg1, arg2, arg3, arg4) => {
         static: {}
     };
     Class._.singleInstance.clear = () => { }; // no operation
+
+    // register type with package
+    flair.Package(Class);
 
     // return
     return Class;